// server/src/types/socketEvents.enum.ts
export enum SocketEventsEnum {
  boardsJoin = "boards:join",
  boardsLeave = "boards:leave",
  boardsUpdate = "boards:update",
  boardsUpdateSuccess = "boards:updateSuccess",
  boardsUpdateFailure = "boards:updateFailure",
  boardsDelete = "boards:delete",
  boardsDeleteSuccess = "boards:deleteSuccess",
  boardsDeleteFailure = "boards:deleteFailure",
  columnCreate = "column:create",
  columnCreateSuccess = "column:createSuccess",
  columnCreateFailure = "column:createFailure",
  columnUpdate = "column:update",
  columnUpdateSuccess = "column:updateSuccess",
  columnUpdateFailure = "column:updateFailure",
  columnDelete = "column:delete",
  columnDeleteSuccess = "column:deleteSuccess",
  columnDeleteFailure = "column:deleteFailure",
  taskCreate = "task:create",
  taskCreateSuccess = "task:createSuccess",
  taskCreateFailure = "task:createFailure",
  taskUpdate = "task:update",
  taskUpdateSuccess = "task:updateSuccess",
  taskUpdateFailure = "task:updateFailure",
  taskDelete = "task:delete",
  taskDeleteSuccess = "task:deleteSuccess",
  taskDeleteFailure = "task:deleteFailure",
}